Moved all the Tiptap extension manifests into a combined/single manifest
