+++ src/packages/rte/tiptap/extensions/core/bullet-list.extension.ts
@@ -1,21 +1,6 @@
 import { UmbTiptapToolbarElementApiBase } from '../types.js';
-import type { ManifestTiptapExtensionButtonKind } from '../tiptap-extension.js';
 import { BulletList, ListItem } from '@umbraco-cms/backoffice/external/tiptap';
 import type { Editor } from '@umbraco-cms/backoffice/external/tiptap';
-
-export const manifest: ManifestTiptapExtensionButtonKind = {
-	type: 'tiptapExtension',
-	kind: 'button',
-	alias: 'Umb.Tiptap.BulletList',
-	name: 'Bullet List Tiptap Extension',
-	api: () => import('./bullet-list.extension.js'),
-	weight: 993,
-	meta: {
-		alias: 'bulletList',
-		icon: 'icon-bulleted-list',
-		label: 'Bullet List',
-	},
-};
 
 export default class UmbTiptapBulletListExtensionApi extends UmbTiptapToolbarElementApiBase {
 	getTiptapExtensions = () => [BulletList, ListItem];
