+++ src/packages/rte/tiptap/extensions/core/heading3.extension.ts
@@ -1,21 +1,6 @@
 import { UmbTiptapToolbarElementApiBase } from '../types.js';
-import type { ManifestTiptapExtensionButtonKind } from '../tiptap-extension.js';
 import { Heading } from '@umbraco-cms/backoffice/external/tiptap';
 import type { Editor } from '@umbraco-cms/backoffice/external/tiptap';
-
-export const manifest: ManifestTiptapExtensionButtonKind = {
-	type: 'tiptapExtension',
-	kind: 'button',
-	alias: 'Umb.Tiptap.Heading3',
-	name: 'Heading 3 Tiptap Extension',
-	api: () => import('./heading3.extension.js'),
-	weight: 947,
-	meta: {
-		alias: 'heading3',
-		icon: 'icon-heading-3',
-		label: 'Heading 3',
-	},
-};
 
 export default class UmbTiptapHeading3ExtensionApi extends UmbTiptapToolbarElementApiBase {
 	getTiptapExtensions = () => [Heading];
