+++ src/packages/rte/tiptap/extensions/core/text-align-left.extension.ts
@@ -1,21 +1,6 @@
 import { UmbTiptapToolbarElementApiBase } from '../types.js';
-import type { ManifestTiptapExtensionButtonKind } from '../tiptap-extension.js';
 import { TextAlign } from '@umbraco-cms/backoffice/external/tiptap';
 import type { Editor } from '@umbraco-cms/backoffice/external/tiptap';
-
-export const manifest: ManifestTiptapExtensionButtonKind = {
-	type: 'tiptapExtension',
-	kind: 'button',
-	alias: 'Umb.Tiptap.TextAlignLeft',
-	name: 'Text Align Left Tiptap Extension',
-	api: () => import('./text-align-left.extension.js'),
-	weight: 919,
-	meta: {
-		alias: 'text-align-left',
-		icon: 'icon-text-align-left',
-		label: 'Text Align Left',
-	},
-};
 
 export default class UmbTiptapTextAlignLeftExtensionApi extends UmbTiptapToolbarElementApiBase {
 	getTiptapExtensions = () => [
